Add Dict implementation based on Compare
Core's dict only tages `comparable` keys. `allDict` takes any key, but
requires you to supply a `key -> comparable` function which might not be
so easy to construct for any given type. This Dict implementation
requires you to supply a `Comparator a` (so `a -> a -> Order`), which is
the most flexible type. If you're already using a `Comparator` in your
code, it should be really easy to get going with this Dict. If you have
a `key -> comparable` style function then this library is already
providing a way to turn that into a comparator.

This copies over most of the code from the allDict implementation, which
in turn seems to take a lot of cueues from the core implementation.